Local Spark Provider which supports to submit feature join job in local spark (#644)
* local spark feature join job with local file

* update local spark with udf support

* add feature gen support in local spark

* update test case

* remove unused feature conf, update doc

* expose master as input and refine local spark provider

+++ feathr_project/feathr/client.py
@@ -2,26 +2,22 @@ import base64
 import logging
 import os
 import tempfile
-from datetime import datetime, timedelta
-from pathlib import Path
-from typing import Dict, List, Optional, Union
-
-from numpy import isin
+from typing import Dict, List, Union
 from feathr.definition.feature import FeatureBase
 
 import redis
 from azure.identity import DefaultAzureCredential
 from jinja2 import Template
 from pyhocon import ConfigFactory
-from feathr.definition.sink import GenericSink, Sink
-from feathr.definition.source import GenericSource, Source
+from feathr.definition.sink import Sink
 from feathr.registry.feature_registry import default_registry_client
 
 from feathr.spark_provider._databricks_submission import _FeathrDatabricksJobLauncher
+from feathr.spark_provider._synapse_submission import _FeathrSynapseJobLauncher
+from feathr.spark_provider._localspark_submission import _FeathrDLocalSparkJobLauncher
 
 from feathr.definition._materialization_utils import _to_materialization_config
 from feathr.udf._preprocessing_pyudf_manager import _PreprocessingPyudfManager
-from feathr.spark_provider._synapse_submission import _FeathrSynapseJobLauncher
 from feathr.constants import *
 from feathr.spark_provider.feathr_configurations import SparkExecutionConfiguration
 from feathr.definition.feature_derivations import DerivedFeature
@@ -36,54 +32,7 @@ from feathr.spark_provider.feathr_configurations import SparkExecutionConfigurat
 from feathr.utils._envvariableutil import _EnvVaraibleUtil
 from feathr.utils._file_utils import write_to_file
 from feathr.utils.feature_printer import FeaturePrinter
-
-
-class FeatureJoinJobParams:
-    """Parameters related to feature join job.
-
-    Attributes:
-        join_config_path: Path to the join config.
-        observation_path: Absolute path in Cloud to the observation data path.
-        feature_config: Path to the features config.
-        job_output_path: Absolute path in Cloud that you want your output data to be in.
-    """
-
-    def __init__(self, join_config_path, observation_path, feature_config, job_output_path, secrets:List[str]=[]):
-        self.secrets = secrets
-        self.join_config_path = join_config_path
-        if isinstance(observation_path, str):
-            self.observation_path = observation_path
-        elif isinstance(observation_path, Source):
-            self.observation_path = observation_path.to_argument()
-            if hasattr(observation_path, "get_required_properties"):
-                self.secrets.extend(observation_path.get_required_properties())
-        else:
-            raise TypeError("observation_path must be a string or a Sink")
-        self.feature_config = feature_config
-        if isinstance(job_output_path, str):
-            self.job_output_path = job_output_path
-        elif isinstance(job_output_path, Sink):
-            self.job_output_path = job_output_path.to_argument()
-            if hasattr(job_output_path, "get_required_properties"):
-                self.secrets.extend(job_output_path.get_required_properties())
-        else:
-            raise TypeError("job_output_path must be a string or a Sink")
-
-
-class FeatureGenerationJobParams:
-    """Parameters related to feature generation job.
-
-    Attributes:
-        generation_config_path: Path to the feature generation config.
-        feature_config: Path to the features config.
-        secrets: secret names from data sources, the values will be taken from env or KeyVault
-    """
-
-    def __init__(self, generation_config_path, feature_config, secrets=[]):
-        self.generation_config_path = generation_config_path
-        self.feature_config = feature_config
-        self.secrets = secrets
-
+from feathr.utils.spark_job_params import FeatureJoinJobParams, FeatureGenerationJobParams
 
 
 class FeathrClient(object):
@@ -162,7 +111,7 @@ class FeathrClient(object):
             'spark_config', 'spark_cluster')
 
         self.credential = credential
-        if self.spark_runtime not in {'azure_synapse', 'databricks'}:
+        if self.spark_runtime not in {'azure_synapse', 'databricks', 'local'}:
             raise RuntimeError(
                 'Only \'azure_synapse\' and \'databricks\' are currently supported.')
         elif self.spark_runtime == 'azure_synapse':
@@ -207,6 +156,14 @@ class FeathrClient(object):
                 databricks_work_dir=self.envutils.get_environment_variable_with_default(
                     'spark_config', 'databricks', 'work_dir')
             )
+        elif self.spark_runtime == 'local':
+            self._FEATHR_JOB_JAR_PATH = \
+                self.envutils.get_environment_variable_with_default(
+                    'spark_config', 'local', 'feathr_runtime_location')
+            self.feathr_spark_launcher = _FeathrDLocalSparkJobLauncher(
+                workspace_path = self.envutils.get_environment_variable_with_default('spark_config', 'local', 'workspace'),
+                master = self.envutils.get_environment_variable_with_default('spark_config', 'local', 'master')
+                )
 
         self._construct_redis_client()
 
@@ -458,6 +415,7 @@ class FeathrClient(object):
                              feature_query: Union[FeatureQuery, List[FeatureQuery]],
                              output_path: Union[str, Sink],
                              execution_configurations: Union[SparkExecutionConfiguration ,Dict[str,str]] = {},
+                             config_file_name:str = "feature_join_conf/feature_join.conf",
                              udf_files = None,
                              verbose: bool = False
                              ):
@@ -468,6 +426,7 @@ class FeathrClient(object):
             feature_query: features that are requested to add onto the observation data
             output_path: output path of job, i.e. the observation data with features attached.
             execution_configurations: a dict that will be passed to spark job when the job starts up, i.e. the "spark configurations". Note that not all of the configuration will be honored since some of the configurations are managed by the Spark platform, such as Databricks or Azure Synapse. Refer to the [spark documentation](https://spark.apache.org/docs/latest/configuration.html) for a complete list of spark configurations.
+            config_file_name: the name of the config file that will be passed to the spark job. The config file is used to configure the spark job. The default value is "feature_join_conf/feature_join.conf".
         """
         feature_queries = feature_query if isinstance(feature_query, List) else [feature_query]
         feature_names = []
@@ -488,7 +447,6 @@ class FeathrClient(object):
             outputPath: "{{output_path}}"
         """)
         config = tm.render(feature_lists=feature_queries, observation_settings=observation_settings, output_path=output_path)
-        config_file_name = "feature_join_conf/feature_join.conf"
         config_file_path = os.path.join(self.local_workspace_dir, config_file_name)
 
         # make sure `FeathrClient.build_features()` is called before getting offline features/materialize features
@@ -542,7 +500,7 @@ class FeathrClient(object):
             main_jar_path=self._FEATHR_JOB_JAR_PATH,
             python_files=cloud_udf_paths,
             job_tags=job_tags,
-            main_class_name='com.linkedin.feathr.offline.job.FeatureJoinJob',
+            main_class_name=JOIN_CLASS_NAME,
             arguments= [
                 '--join-config', self.feathr_spark_launcher.upload_or_get_cloud_path(
                     feature_join_job_params.join_config_path),
@@ -623,7 +581,7 @@ class FeathrClient(object):
         for sink in settings.sinks:
             if hasattr(sink, "get_required_properties"):
                 secrets.extend(sink.get_required_properties())
-        
+        results = []
         # produce materialization config
         for end in settings.get_backfill_cutoff_time():
             settings.backfill_time.end = end
@@ -642,13 +600,16 @@ class FeathrClient(object):
 
             udf_files = _PreprocessingPyudfManager.prepare_pyspark_udf_files(settings.feature_names, self.local_workspace_dir)
             # CLI will directly call this so the experience won't be broken
-            self._materialize_features_with_config(config_file_path, execution_configurations, udf_files, secrets)
-            if os.path.exists(config_file_path):
+            result = self._materialize_features_with_config(config_file_path, execution_configurations, udf_files, secrets)
+            if os.path.exists(config_file_path) and self.spark_runtime != 'local':
                 os.remove(config_file_path)
+            results.append(result)
 
         # Pretty print feature_names of materialized features
         if verbose and settings:
             FeaturePrinter.pretty_print_materialize_features(settings)
+
+        return results
 
     def _materialize_features_with_config(self, feature_gen_conf_path: str = 'feature_gen_conf/feature_gen.conf',execution_configurations: Dict[str,str] = {}, udf_files=[], secrets=[]):
         """Materializes feature data based on the feature generation config. The feature
@@ -688,7 +649,7 @@ class FeathrClient(object):
             job_name=self.project_name + '_feathr_feature_materialization_job',
             main_jar_path=self._FEATHR_JOB_JAR_PATH,
             python_files=cloud_udf_paths,
-            main_class_name='com.linkedin.feathr.offline.job.FeatureGenJob',
+            main_class_name=GEN_CLASS_NAME,
             arguments=arguments,
             reference_files_path=[],
             configuration=execution_configurations,
@@ -717,7 +678,7 @@ class FeathrClient(object):
         REDIS_PORT: {REDIS_PORT}
         REDIS_SSL_ENABLED: {REDIS_SSL_ENABLED}
         """.format(REDIS_PASSWORD=password, REDIS_HOST=host, REDIS_PORT=port, REDIS_SSL_ENABLED=ssl_enabled)
-        return config_str
+        return self._reshape_config_str(config_str)
 
     def _get_s3_config_str(self):
         """Construct the S3 config string. The endpoint, access key, secret key, and other parameters can be set via
@@ -733,7 +694,7 @@ class FeathrClient(object):
             S3_ACCESS_KEY: "{S3_ACCESS_KEY}"
             S3_SECRET_KEY: "{S3_SECRET_KEY}"
             """.format(S3_ENDPOINT=endpoint, S3_ACCESS_KEY=access_key, S3_SECRET_KEY=secret_key)
-        return config_str
+        return self._reshape_config_str(config_str)
 
     def _get_adls_config_str(self):
         """Construct the ADLS config string for abfs(s). The Account, access key and other parameters can be set via
@@ -747,7 +708,7 @@ class FeathrClient(object):
             ADLS_ACCOUNT: {ADLS_ACCOUNT}
             ADLS_KEY: "{ADLS_KEY}"
             """.format(ADLS_ACCOUNT=account, ADLS_KEY=key)
-        return config_str
+        return self._reshape_config_str(config_str)
 
     def _get_blob_config_str(self):
         """Construct the Blob config string for wasb(s). The Account, access key and other parameters can be set via
@@ -761,7 +722,7 @@ class FeathrClient(object):
             BLOB_ACCOUNT: {BLOB_ACCOUNT}
             BLOB_KEY: "{BLOB_KEY}"
             """.format(BLOB_ACCOUNT=account, BLOB_KEY=key)
-        return config_str
+        return self._reshape_config_str(config_str)
 
     def _get_sql_config_str(self):
         """Construct the SQL config string for jdbc. The dbtable (query), user, password and other parameters can be set via
@@ -781,7 +742,7 @@ class FeathrClient(object):
             JDBC_AUTH_FLAG: {JDBC_AUTH_FLAG}
             JDBC_TOKEN: {JDBC_TOKEN}
             """.format(JDBC_TABLE=table, JDBC_USER=user, JDBC_PASSWORD=password, JDBC_DRIVER = driver, JDBC_AUTH_FLAG = auth_flag, JDBC_TOKEN = token)
-        return config_str
+        return self._reshape_config_str(config_str)
 
     def _get_monitoring_config_str(self):
         """Construct monitoring-related config string."""
@@ -795,7 +756,7 @@ class FeathrClient(object):
                 MONITORING_DATABASE_SQL_USER: {user}
                 MONITORING_DATABASE_SQL_PASSWORD: {password}
                 """.format(url=url, user=user, password=password)
-            return config_str
+            return self._reshape_config_str(config_str)
         else:
             ""
 
@@ -813,7 +774,7 @@ class FeathrClient(object):
             JDBC_SF_ROLE: {JDBC_SF_ROLE}
             JDBC_SF_PASSWORD: {JDBC_SF_PASSWORD}
             """.format(JDBC_SF_URL=sf_url, JDBC_SF_USER=sf_user, JDBC_SF_PASSWORD=sf_password, JDBC_SF_ROLE=sf_role)
-        return config_str
+        return self._reshape_config_str(config_str)
 
     def _get_kafka_config_str(self):
         """Construct the Kafka config string. The endpoint, access key, secret key, and other parameters can be set via
@@ -823,7 +784,7 @@ class FeathrClient(object):
         config_str = """
             KAFKA_SASL_JAAS_CONFIG: "{sasl}"
             """.format(sasl=sasl)
-        return config_str
+        return self._reshape_config_str(config_str)
 
     def _collect_secrets(self, additional_secrets=[]):
         """Collect all values corresponding to the secret names."""
@@ -846,4 +807,10 @@ class FeathrClient(object):
                 feature_dict[feature.name] = feature
         for feature in registry_derived_feature_list:
                 feature_dict[feature.name] = feature
-        return feature_dict
+        return feature_dict 
+
+    def _reshape_config_str(self, config_str:str):
+        if self.spark_runtime == 'local':
+            return "'{" + config_str + "}'"
+        else:
+            return config_str
